Get rid of psutil/helpers.py namespace
