+++ scripts/nettop.py
@@ -39,7 +39,7 @@ except ImportError:
     sys.exit('platform not supported')
 
 import psutil
-from psutil.helpers import bytes2human
+from psutil import bytes2human
 
 lineno = 0
 win = curses.initscr()
