+++ scripts/pmap.py
@@ -32,7 +32,7 @@ import shutil
 import sys
 
 import psutil
-from psutil.helpers import bytes2human
+from psutil import bytes2human
 
 
 def safe_print(s):
